[IMP] show the removal date in transfers
Of the four dates recorded on prodlots, the removal date is the most meaninful in the generic case, because this is the date that the scheduler uses to enforce FEFO warehouse rules.
The Life date on the other hand is only meaningful in indistries with very strict sanitary rules. Many industries even in agro would not use us but use best-before date instead.